cant use float32 now...

+++ settings.py
@@ -1,2 +1,2 @@
-# FLOAT_TYPE = 'float64'
-FLOAT_TYPE = "float32"
+FLOAT_TYPE = "float64"
+# FLOAT_TYPE = "float32"
